Feat: 사물함까지 선택해야  사물함 정보 표시#26
