+++ src/pages/SideNavigationLayout.tsx
@@ -10,6 +10,7 @@ interface NavBuildingProps {
   selectedBuilding: string | null; // 선택된 건물의 인덱스 또는 null
   setSelectedBuilding: (name: string | null) => void; // 선택된 건물을 설정하는 함수
   setSelectedFloor: (floor: number | null) => void; // 선택된 층을 설정하는 함수
+  setSelectedCabinet: (cabinet: number | null) => void; // 추가 (드롭다운에서 건물 선택 시 사물함 정보도 초기화)
 }
 
 const SideNavigationLayout = ({
@@ -17,6 +18,7 @@ const SideNavigationLayout = ({
   selectedBuilding,
   setSelectedBuilding,
   setSelectedFloor,
+  setSelectedCabinet,
 }: NavBuildingProps) => {
   const { isOpen, setIsOpen, dropdownOutsideRef } = useBuildingState();
   const location = useLocation();
@@ -72,6 +74,7 @@ const SideNavigationLayout = ({
                     onClick={() => {
                       setSelectedBuilding(building.name); // 선택한 건물 업데이트
                       setSelectedFloor(null); // 건물 층수 초기화
+                      setSelectedCabinet(null);
                       setIsOpen(false); // 드롭다운 닫기
                     }}
                   >
